#69 rename: 파일명 제대로 변경, getProfile 파일 경로 수정
