+++ components/wikiEdit/WikiEditForm.tsx
@@ -4,9 +4,9 @@ import { FormEvent } from 'react';
 import { useState, useEffect } from 'react';
 import { patchUser } from '@/apis/auth/patchUser';
 import { useRouter } from 'next/router';
-import { getProfile } from '@/apis/auth/getProfile';
+import { getProfile } from '@/apis/getProfile';
 import WikiEditProfile from './WikiEditProfile';
-import styles from './WIkiEditForm.module.scss';
+import styles from './WikiEditForm.module.scss';
 import { FormValue } from './types/EditTypes';
 import { useAuth } from '@/contexts/AuthProvider';
 import ModalComponent from '../@shared/modal/Modal';
